+++ NEWS.md
@@ -1,3 +1,5 @@
+# hpa.spatial (development version)
+
 # hpa.spatial 0.2.0
 
 * create an `lhn` dataset to replace the independent state datasets (`qld_hhs`,