+++ NEWS.md
@@ -1,4 +1,4 @@
-# hpa.spatial (development version)
+# hpa.spatial 0.2.1
 
 * add `get_correspondence_tbl()` which calls `strayr::read_correspondence_tbl()`
   and if it fails to find one, it uses `make_correspondence_tbl()` to create